Set DefaultPowerReduction in app.go, added flags to script, changed swap calculation, made backup folder unique and added useful swap scripts.

+++ genesisd.sh
@@ -32,17 +32,120 @@ THIS IS AN UPGRADE TO COSMOS SDK V0.46.15 BASED ON CRONOS RELEASE SOURCE CODE, T
   
 EOF
 
-moniker=""
+# User-configurable variables
 minimum_combined_gb=150
-repo_dir=$(cd "$(dirname "$0")" && pwd)
-total_ram_kb=$(grep MemTotal /proc/meminfo | awk '{print $2}')
-total_swap_kb=$(grep SwapTotal /proc/meminfo | awk '{print $2}')
-total_combined_gb=$((($total_ram_kb + $total_swap_kb) / 1024 / 1024))
-available_disk_gb=$(df -BG --output=avail / | awk 'NR==2 {print $1}' | tr -d 'G')
 disk_headroom_gb=50
+backup_dir=".genesisd_backup_$(date +"%Y%m%d%H%M%S")"
 
-# Function to add a line to a file if it doesn't already exist (to prevent duplicates)
-# Usage: add_line_to_file "line" file [use_sudo]
+# Fixed/default variables (do not modify)
+moniker=""
+repo_dir=$(cd "$(dirname "$0")" && pwd)
+crisis_skip=false
+skip_state_download=false
+reset_priv_val_state=false
+no_service=false
+no_start=false
+
+if [ "$(id -u)" -ne 0 ]; then
+    echo "This script must be run as the root user."
+    exit 1
+fi
+
+if [ "$#" -lt 1 ]; then
+    echo "Usage: $0 <command> [moniker]"
+    echo "   <command> should be either 'upgrade' or 'init'"
+    echo ""
+    echo "   Options:"
+    echo "     --crisis-skip            Makes sure that genesisd starts with the --x-crisis-skip-assert-invariants flag (default: false)"
+    echo "     --skip-state-download    Skips downloading the genesis.json file, only do this if you're certain to have the correct state file already (default: false)"
+    echo "     --reset-priv-val-state   Resets data/priv_validator_state.json file [UNSAFE] (default: false)"
+    echo "     --no-service             This prevents the genesisd service from being made (default: false)"
+    echo "     --no-start               This prevents the genesisd service from starting at the end of the script (default: false)"
+    exit 1
+fi
+
+case "$1" in
+    "upgrade")
+        if [ "$#" -eq 1 ] || { [ "$#" -ge 2 ] && [ "$(echo "$2" | cut -c 1-2)" = "--" ]; }; then
+            moniker=$(grep "moniker" ~/.genesisd/config/config.toml | cut -d'=' -f2 | tr -d '[:space:]"')
+            
+            if [ -z "$moniker" ]; then
+                echo "Error: No moniker found in the current configuration nor has one been provided as an argument."
+                exit 1
+            fi
+            
+            echo "Upgrade mode with moniker from previous configuration: $moniker"
+        elif [ "$#" -ge 2 ]; then
+            moniker="$2"
+            echo "Upgrade mode with moniker: $moniker"
+        else
+            echo "Invalid number of arguments for 'upgrade' mode. Usage: $0 upgrade [moniker]"
+            exit 1
+        fi
+        ;;
+    "init")
+        if [ "$#" -ge 2 ]; then
+            if [ "$(echo "$2" | cut -c 1-2)" = "--" ]; then
+              echo "Missing or invalid argument for 'init' mode. Usage: $0 init <moniker>"
+              exit 1
+            fi
+
+            moniker="$2"
+            echo "Init mode with moniker: $moniker"
+        else
+            echo "Missing or invalid argument for 'init' mode. Usage: $0 init <moniker>"
+            exit 1
+        fi
+        ;;
+    *)
+        echo "Invalid command: $1. Please use 'upgrade' or 'init'."
+        exit 1
+        ;;
+esac
+
+# Process command-line arguments and set corresponding flags
+for arg in "$@"; do
+    case "$arg" in
+        --crisis-skip)
+            crisis_skip=true
+            ;;
+        --skip-state-download)
+            skip_state_download=true
+            ;;
+        --reset-priv-val-state)
+            reset_priv_val_state=true
+            ;;
+        --no-service)
+            no_service=true
+            ;;
+        --no-start)
+            no_start=true
+            ;;
+        *)
+            # Handle other arguments or flags here
+            ;;
+    esac
+done
+
+$crisis_skip && echo "o Will add the '--x-crisis-skip-assert-invariants'-flag to the genesisd.service (--crisis-skip: $crisis_skip)"
+$skip_state_download && echo "o Will skip downloading the genesis.json file (--skip-state-download: $skip_state_download)"
+$reset_priv_val_state && echo "o Will reset the data/priv_validator_state.json file [UNSAFE] (--reset-priv-val-state: $reset_priv_val_state)"
+! $reset_priv_val_state && echo "o Will preserve the data/priv_validator_state.json (--reset-priv-val-state: $reset_priv_val_state)"
+$no_service && echo "o Will skip installing genesisd as a service (--no-service: $no_service)"
+if ! $no_service && $no_start; then
+    echo "o Will skip starting the genesisd service at the end of the script (--no-start: $no_start)"
+fi
+
+echo ""
+echo "Please note that the Genesis daemon will be halted before proceeding. You will have a 20-second window to cancel this action."
+sleep 20s
+
+service genesis stop
+service genesisd stop
+
+sleep 3s
+
+# Function: adds a line to a file if it doesn't already exist (to prevent duplicates)
 add_line_to_file() {
     local line="$1"
     local file="$2"
@@ -61,52 +164,12 @@ add_line_to_file() {
     fi
 }
 
-if [ "$(id -u)" -ne 0 ]; then
-    echo "This script must be run as the root user."
-    exit 1
-fi
-
-if [ "$#" -lt 1 ]; then
-    echo "Usage: $0 <command> [moniker]"
-    echo "   <command> should be either 'upgrade' or 'init'"
-    exit 1
-fi
-
-case "$1" in
-    "upgrade")
-        if [ "$#" -eq 1 ]; then
-            moniker=$(grep "moniker" ~/.genesisd/config/config.toml | cut -d'=' -f2 | tr -d '[:space:]"')
-            
-            if [ -z "$moniker" ]; then
-                echo "Error: No moniker found in the current configuration nor has one been provided as an argument."
-                exit 1
-            fi
-            
-            echo "Upgrade mode with moniker from previous configuration: $moniker"
-        elif [ "$#" -eq 2 ]; then
-            moniker="$2"
-            echo "Upgrade mode with moniker: $moniker"
-        else
-            echo "Invalid number of arguments for 'upgrade' mode. Usage: $0 upgrade [moniker]"
-            exit 1
-        fi
-        ;;
-    "init")
-        if [ "$#" -eq 2 ]; then
-            moniker="$2"
-            echo "Init mode with moniker: $moniker"
-        else
-            echo "Missing or invalid argument for 'init' mode. Usage: $0 init <moniker>"
-            exit 1
-        fi
-        ;;
-    *)
-        echo "Invalid command: $1. Please use 'upgrade' or 'init'."
-        exit 1
-        ;;
-esac
-
 # ADD ADDITIONAL SWAP (IF NECESSARY)
+total_ram_kb=$(grep MemAvailable /proc/meminfo | awk '{print $2}')
+total_swap_kb=$(grep SwapFree /proc/meminfo | awk '{print $2}')
+total_combined_gb=$((($total_ram_kb + $total_swap_kb) / 1024 / 1024))
+available_disk_gb=$(df -BG --output=avail / | awk 'NR==2 {print $1}' | tr -d 'G')
+
 if [ "$total_combined_gb" -lt "$minimum_combined_gb" ]; then
     # Calculate additional swap space needed in gigabytes
     additional_swap_gb=$((minimum_combined_gb - total_combined_gb + 1)) 
@@ -143,7 +206,7 @@ else
     echo "No additional swap space needed."
 fi
 
-sleep 15s
+sleep 3s
 
 # SYSTEM UPDATE, INSTALLATION OF THE FOLLOWING PACKAGES: jq git wget make gcc build-essential snapd wget ponysay, INSTALLATION OF GO 1.20 via snap
 
@@ -165,20 +228,22 @@ ulimit -n 50000
 snap install ponysay
 ponysay "Installing genesisd from source code with updated genesis_29-2 mainnet!"
 sleep 5s
-ponysay "WARNING: cosmosvisor, evmosd processes will be killed and genesis, genesisd, evmos, evmosd system services will be stopped with this script on the next step. If you have other blockchains running, you might want to delete those parts of the script!"
+ponysay "WARNING: cosmosvisor, evmosd processes will be killed and evmos, evmosd system services will be stopped with this script on the next step. If you have other blockchains running, you might want to delete those parts of the script!"
 sleep 20s
 
 #STOPPING EVMOSD DAEMON AND COSMOVISOR IF IT WAS NOT STOPPED
 pkill evmosd
 pkill cosmovisor
-service genesis stop
-service genesisd stop
 service evmos stop
 service evmosd stop
 
 # BACKUP genesis_29-2 (evmos version) .genesisd
 cd
-rsync -r --verbose --exclude 'data' ./.genesisd/ ./.genesisd_backup/
+rsync -r --verbose --exclude 'data' ./.genesisd/ ./"$backup_dir"/
+mkdir -p ./"$backup_dir"/data
+if cp ./.genesisd/data/priv_validator_state.json ./"$backup_dir"/data/priv_validator_state.json; then
+    echo "Backed up priv_validator_state.json file"
+fi
 
 # DELETING OF .genesisd FOLDER (PREVIOUS INSTALLATIONS)
 cd
@@ -191,7 +256,7 @@ make install
 
 # COPY .genesisd_backup FOLDER to .genesisd FOLDER, EXCLUDE data
 cd
-rsync -r --verbose --exclude 'data' ./.genesisd_backup/ ./.genesisd/
+rsync -r --verbose --exclude 'data' ./"$backup_dir"/ ./.genesisd/
 
 # SETTING UP THE NEW chain-id in CONFIG
 genesisd config chain-id genesis_29-2
@@ -213,14 +278,22 @@ if [ "$1" = "init" ]; then
 fi
 
 #IMPORTING GENESIS STATE
-cd 
-cd .genesisd/config
-rm -r genesis.json
-wget http://135.181.135.29/genesisd/genesis.json
+if ! $skip_state_download; then
+    cd 
+    cd .genesisd/config
+    rm -r genesis.json
+    wget http://135.181.135.29/genesisd/genesis.json
+fi
 cd
 
 # RESET TO IMPORTED genesis.json
 genesisd tendermint unsafe-reset-all
+
+if ! $reset_priv_val_state; then
+    if cp ./"$backup_dir"/data/priv_validator_state.json ./.genesisd/data/priv_validator_state.json; then
+        echo "Restored backed up priv_validator_state.json file"
+    fi
+fi
 
 # CONFIG FILES
 cd ~/.genesisd/config
@@ -234,13 +307,19 @@ sed -i "s/moniker = \"\"/moniker = \"$moniker\"/" config.toml
 echo "Moniker value set to: $moniker"
 
 # SETTING genesisd AS A SYSTEMD SERVICE
-sudo cp "$repo_dir/genesisd.service" /etc/systemd/system/genesisd.service
-systemctl daemon-reload
-systemctl enable genesisd
-# echo "All set!" 
-sleep 3s
+if ! $no_service; then
+    if $crisis_skip; then
+        sudo cp "$repo_dir/genesisd-crisis.service" /etc/systemd/system/genesisd.service
+    else 
+        sudo cp "$repo_dir/genesisd.service" /etc/systemd/system/genesisd.service
+    fi
 
-# STARTING NODE
+    systemctl daemon-reload
+    systemctl enable genesisd
+    sleep 3s
+
+    # STARTING NODE
+    if ! $no_start; then
 cat << "EOF"
      	    \\
              \\_
@@ -249,8 +328,12 @@ cat << "EOF"
        Node start                                                                                                                                                                                     
 EOF
  
-sleep 5s
-systemctl start genesisd
-
-# genesisd start
-ponysay "genesisd node service started, you may try *journalctl -fu genesisd -ocat* or *service genesisd status* command to see it! Welcome to GenesisL1 blockchain!"
+        sleep 5s
+        systemctl start genesisd
+        ponysay "genesisd node service started, you may try *journalctl -fu genesisd -ocat* or *service genesisd status* command to see it! Welcome to GenesisL1 blockchain!"
+    else
+        ponysay "genesisd node service installed, use *service genesisd start* to start it! Welcome to GenesisL1 blockchain!"
+    fi
+else
+    ponysay "genesisd node is ready, use *service genesisd start* to start it! Welcome to GenesisL1 blockchain!"
+fi
